requirements.txt
sanic>=21.6
sqlalchemy>=1.4
asyncpg
httpx
PyCryptodome
cryptography
aioredis==1.3.1
marshmallow_sqlalchemy==0.26.1
